테스트 커버리지 수집을 위한 폴더 구조 변경
