+++ jest.config.js
@@ -114,7 +114,7 @@ module.exports = {
   // restoreMocks: false,
 
   // The root directory that Jest should scan for tests and modules within
-  rootDir: "test",
+  // rootDir: "",
 
   // A list of paths to directories that Jest should use to search for files in
   // roots: [
