+++ utils/apiService/apiService.client.test.ts
@@ -1,11 +1,7 @@
-/**
- * @jest-environment node
- */
-
 jest.mock("axios");
 
 import axios from "axios";
-import apiSerivce from "../../utils/apiService";
+import apiSerivce from "./apiService";
 
 const url = "/api/request";
 
@@ -15,9 +11,8 @@ describe("ApiSerivce", () => {
     axios.get.mockReturnValue(Promise.resolve(1));
 
     const data = await apiSerivce.get(url);
-
     expect(axios.get).toHaveBeenCalledTimes(1);
-    expect(axios.get).toHaveBeenCalledWith(`http://localhost:3000${url}`);
+    expect(axios.get).toHaveBeenCalledWith(url);
     expect(data).toBe(1);
   });
   test("should be request with proper url using POST", async () => {
@@ -26,9 +21,7 @@ describe("ApiSerivce", () => {
 
     const data = await apiSerivce.post(url, { name: "john" });
     expect(axios.post).toHaveBeenCalledTimes(1);
-    expect(axios.post).toHaveBeenCalledWith(`http://localhost:3000${url}`, {
-      name: "john"
-    });
+    expect(axios.post).toHaveBeenCalledWith(url, { name: "john" });
     expect(data).toEqual({ age: 27 });
   });
   test("should be request with proper url using PUT", async () => {
@@ -37,7 +30,7 @@ describe("ApiSerivce", () => {
 
     const data = await apiSerivce.put(url);
     expect(axios.put).toHaveBeenCalledTimes(1);
-    expect(axios.put).toHaveBeenCalledWith(`http://localhost:3000${url}`);
+    expect(axios.put).toHaveBeenCalledWith(url);
     expect(data).toEqual({ success: true });
   });
   test("should be request with proper url using DELETE", async () => {
@@ -47,7 +40,7 @@ describe("ApiSerivce", () => {
 
     const data = await apiSerivce.delete(url);
     expect(axios.delete).toHaveBeenCalledTimes(1);
-    expect(axios.delete).toHaveBeenCalledWith(`http://localhost:3000${url}`);
+    expect(axios.delete).toHaveBeenCalledWith(url);
     expect(data).toBe(returnValue);
   });
 });
