+++ utils/apiService/index.ts
@@ -1,0 +1,3 @@
+import apiService from "./apiService";
+
+export default apiService;
