+++ utils/getCurrentUrl/index.ts
@@ -1,0 +1,2 @@
+import getCurrentUrl from "./getCurrentUrl";
+export default getCurrentUrl;
